% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/judge_chunk_code.R
\name{judge_chunk_code}
\alias{judge_chunk_code}
\title{Title}
\usage{
judge_chunk_code(chunk_name, judgement = "you wrote code", ...)
}
\arguments{
\item{chunk_name}{}

\item{judgement}{}

\item{...}{}
}
\description{
Title
}
