% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/judge_chunk_output_plot.R
\name{judge_chunk_output_plot}
\alias{judge_chunk_output_plot}
\title{Title}
\usage{
judge_chunk_output_plot(chunk_name, judgement = "some output")
}
\arguments{
\item{chunk_name}{}

\item{judgement}{}
}
\description{
Title
}
