% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/judge_chunk_output_text.R
\name{judge_chunk_output_text}
\alias{judge_chunk_output_text}
\title{Title}
\usage{
judge_chunk_output_text(chunk_name, judgement = "some output")
}
\arguments{
\item{chunk_name}{}

\item{judgement}{}
}
\description{
Title
}
