% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/judge_plot.R
\name{judge_plot}
\alias{judge_plot}
\title{Title}
\usage{
judge_plot(
  plot = NULL,
  color = "red",
  alpha = 0.9,
  judgement = "you made\\na plot",
  family = "Helvetica",
  fontface = "bold",
  clip = "off",
  plot.margin = margin(50, 70, 10, 10),
  x = 1,
  y = 1,
  vjust = 1.1,
  hjust = 1.1,
  size = 35,
  angle = 0
)
}
\arguments{
\item{plot}{}

\item{color}{}

\item{alpha}{}

\item{judgement}{}

\item{family}{}

\item{fontface}{}

\item{clip}{}

\item{plot.margin}{}

\item{x}{}

\item{y}{}

\item{vjust}{}

\item{hjust}{}

\item{size}{}

\item{angle}{}
}
\description{
Title
}
